% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSfrags.R
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{MSfrags
Sample peptide fragments in Sprectronaut format}
\format{
A vector for sample peptide fragments with protein modifications
}
\source{
These fragments were generated from P51681 (for details see...)
}
\usage{
MSfrags
}
\description{
MSfrags
Sample peptide fragments in Sprectronaut format
}
\keyword{datasets}
